Crate generate_format

How and where to record the Serde format of interesting Diem types. See API documentation with cargo doc -p serde-reflection --open

Enums

Corpus
A corpus of Rust types to trace, and optionally record on disk.

Functions

lint_bcs_format
Verify that a Serde format is compatible with BCS and follows best practices.